{"type":"TI", "runs":"0"}

+++ Breakout.java
@@ -69,7 +69,7 @@ public class Breakout extends GraphicsProgram {
 	private static final int N_COLOR_ROWS = NBRICK_ROWS/5;
 	private static final int TIME = 10; 
 	private static double brickDif = BRICK_HEIGHT + BRICK_SEP; 
-	private static double SCORE_OFFSET = BRICK_Y_OFFSET / 5;
+	private static int SCORE_OFFSET = BRICK_Y_OFFSET / 5;
 	private static int SCORE_SIZE = SCORE_OFFSET/2; 
 	
 /* Method: run() */
